Advent of Code day 1
Given a bunch of elves, get the elf with the most total
Then get the top 3 elves' totals.

Followed Chris Biscardi to figure out how the project should be
structured.
But then I deviated from it.
If I'm going to do this for multiple days, doesn't it make sense to have
the code for each day in its own file?
Instead of putting code in lib.rs and making a whole new project each day?
I don't know how to structure it, but it works at least