Day 2
Works.
Might be able to make it look better / refactor it.
Reused most of the code from part 1 to do part 2 so maybe that means I
split stuff up nicely?

get_variant
sloppy
Checking for errors in input
get_variant_2 is for the new requirements, but get_variant() still works
as long as you don't pass in the xyz input. So it works to keep using
get_variant, but could break if I passed in the wrong input.