Day 3 refactor
Refactored the matching letter logic to work for an arbitrary amount of
inputs.
Moving badge_priority_total to the top messed up the diff a bit. I just
moved it

put each letter of one input into a set.
Then on the next input, check if this set contains the letter. If it does,
put it into a new set. This second set will be the intersection of the
first 2 inputs. When you do the third input, you can use this second set
to check; you don't need the first set anymore so you can clear it and
insert into it.
There's probably a way to do this switching without duplicating so much
code and having the if use_set_a conditions

The 2 parts are pretty similar
split a line in half and check collisions
take 3 lines and check for collisions