Day 04 Refactor
Part1() and part2() were very similar so I made the shared code its own fn
get_pairs()

get_points_part1
where the real logic of the part 1 is
cargo clippy told me to convert from bool to u32 instead of returning 1 or
0. It kinda hurts readability I think, but oh well.

get_points_part2
might be not super readable. Comments make is clear as mud. Maybe it would
be better to hide this logic in a PairRange.some_overlap()