Day 04 Readability
Made a helper method to check for some overlap, PairRange.some_overlap()
It hides the !(a.less(b) || b.less(a)) logic behind a name that describes
what it does. Makes it more clear what is happening. If you care how to
works, you can check the fn and read the comment there.

as u32 seems more readable to me than u32::from() even though they are
equivalent. Maybe because it sits on its own. There's a gap between the
logic and the converting to a number that separates them and allows you to
think about them independantly.