Day 07 Part 1
Works

Create a tree
Put all the files and folders in the tree.

Making the tree was easy. The hard part was getting parent to work so I
could do cd ..
Had to change the tree to use Rc<RefCell<TreeNode>>

Still has a lot of code from the og implementation that needs to be cut
out.
Tried the strategy of doing the changes in parallel. I kinda did it wrong.
I think I'm supposed to use the same apis with the new arguments so it can
keep the same tests and fns. I made new tests and fns.
Oh well, it was my first time learning Rc so I kinda understand it now.