Day 10 Part 2
Rank 7989

Weird offsets in this. cycles are from 1.. not from 0..
Count also starts at 1.
Made converting 41..80 -> 1..40 weirder. Need (cycle - 1) % 40 + 1 instead
of just cycle % 40.

visual test
prints the correct output then mine under it.
Normal assert(string, string) ignores the formatting so it's hard to see
on larger blocks of text (but does let me catch the \n at the end that
shouldn't be there).
Test passes so it doesn't output. To get the output, run just that test or
remove the assert(true)
Maybe it should be rolled into the one test. If it fails, it will produce
the visual output and the raw output with the \n's.

Is there a way to see where in the string it doesn't match instead of just
dumping both?

for both parts, I copied the cycle += 1 logic 3 times. Maybe this is the
most clear way? or should I do for 0..2 ?