Day 11 Part 2
Rank 7284
Part 1 was 9668 so I gained a lot of rankes part 1 to 2. I do usually gain
some ranks part 1 to 2.

Part 2 is basically the same as part 1
Just need to pass the num rounds (20 -> 10_000) and the worry reduction
(1/3 -> 1/1)

Without the /3 reduction, the values will overflow. To reduce them, % them
by a common factor. To find a common factor, multiply all tests together.
Still an issue bc this factor squared is bigger than a u32. So I needed to
upgrade everything to a usize(u64). usize is platform dependent so maybe
u64 would be better. Can 32-bit machines use u64? is usize just accounting
for preference for certain things, but any machine can still use u64?

Took me a while to figure out the reducing wouldn't help. I wasn't sure if
I wasn't doing it right and that's why it was overflowing. Checking the
factor squared overflows answers that question. If that overflows, then
even if it was working correctly, it would overflow.

worry reducer
could be a field in the Monkey instead of being passed in. Oh well,
doesn't really matter.