Day 12 Part 1
Works

Tests worked because the start was at 0, but were bugged for different
starts.
get_current was wrong. I was setting the default lowest too low. The logic
in it is also maybe backwards. If there is only 1 element in the open_set,
that should be the default instead of the default being 0.

I didn't test this well. I was writing out the algorithm and figuring out
what data I needed then I was going to put it into its own fn so I could
test, but I didn't. whoops.