Day 12 Part 2
Rank 9318

Part 2 diff
Instead of checking if letter == S for the start, check for S and a
Make a vec of all starts. Perform a search for each start. Get the lowest.
get_path_length returns 10_000 if no path found. Should return an option
instead. 10_000 is easy so sorting by smallest still works.

This feels slow to run. I wonder if I'm doing something wrong? Maybe I
could use a heap to sort the open set, but then I'd need to pass in the
f_cost and index as a struct. Are those the only values I need?