Day 13 Part 1 Start
I can't figure this one out.

struct Signal might be the way to convert from text into code.

Depth
Maybe I can count the number of opening and closing[] to figure out where
each number fits in the sequence?
[[1],[2,3,4]] all have depth 2 so I would need to track which list they
are a part of. 1 is part of the 2nd list? 2,3,4 are part of the 3rd? There
would be no numbers part of the 1st list.
[[1],4] 1 is part of the 2nd list, 4 is part of the 1st list?
Maybe this doesn't work either.