Day 13 Failed
Doesn't work

Tests pass, but fail on the real input.

Tried a couple strategies. Couldn't get the recursion to work.
I want List[List[1,2],3,4], but I get List[List[1,2], List[3,4]]

I want to create a list on [ and close the list on ], but I don't know
what that means in code.