Day 15 Part 2
Rank 5298

First version of this ran 4m * 4m * 26 times
To do just 4m * 26 takes 5s
4m * 5s = 5_555 hours

Version 2
For each sensor, find the edge points.
For each edge point, compare with each other sensor
Instead of iterating every item in the whole grid, only iterate the edge
of each sensor.

I have a test to get the intersection of multiple sets, but the iteration
version was easier.
It's a it awkward to coerce a set.intersection(set) back into a set so I
didn't want to do it procedurally.

diamond
grabbed the logic from my tower defense games.
Added the bounds.
The bounds aren't really tested. It works for the test input and the real
input so that's some integration testing at least.

Timing
Kind of figured out some timing.
Might use it on each day going forward to get a rough idea how long these
take.

Code needs to be cleaned up.
Comments/notes are in the way making it hard to read.