Day 16 Part 1 Works
Works

I was only checking if current > best when the depth was reached. If I
check each tie it works. Why?
Where does it return otherwise?

Added a check so you don't visit valves that are too far away to get to.
If you only have 2 min left, can't get to a valve with dist of 7.

I also had time * flow instead of new_time * flow

The combination of these fixes fixed it.
I think I should only need the dist check and new_time instead of time to
make it work, but I needed all 3. Not sure why.