Day 17 Part 1
Works

Made a visualizer too.
Had a bug where I was setting the height to the last block instead of
checking to see if it was the highest block.

Part 2
Need to do a lot of iterations
stuff overflows

I think the trick will be to find when stuff repeats.
Looking through the giant stack part 1 made, it seemed fairly repetitive.