Day 17 Part 2 Start
Doesn't work
I tried to find when a full row is filled. When this happens, I can
deleted everything below it. Reset the values of stuff above the line to
small numbers so the heigh number never gets too big.
Except that isn't really the issue.
The issue is a trillion is too big. I can't simulate that much. a million
takes 25s. A trillion would take like 200 days.

This new culling version doesn't even work on part 1. it's close, but
probably off by 1 somewhere that I can't find.

Back to the drawing board trying to figure out how to find when it
repeats.