Day 18 Part 2
Rank 3200, 3091
Best I've done. And done in <2hrs

3D floodfill
easier to implement than I thought it was going to be.
wikipedia told me how to use a queue instead of making it recursive which
was nice of them.

bit hacky finding the outer faces
test input is 1..3. real input is 0..21
so I figured I could use Cube x: u32
but when I did the air, it made air blocks all around the test input, but
the real input didn't have air blocks at the -1 positions.
Maybe makign Cube {x: i32} would have been a better solution to get the
air placement consistent.
I made neighbours return a vec on none if the neighbour would be out of
bounds.

The real input has the rocks block off part of the bigger cube.
So if you flood fill from 0,0,0, it will block off some stuff that is
accessable to air if you start at 22,22,22 or something.
So the hack is to have the air extend out of bounds in the positive x,y,z
directions so it can get to everything.
It would fail if there were a 'cave' along the x,y,or z face in the
negative direction. Luckily there wasn't bc I just thought of that now and
I don't know if I would've figured that out if the test failed.