Day 20 Part 1
Works
Brute force solution

had an idea to have a vec of the values and a vec of pointers to the
values. So I would have an ordered list (the pointers) to get the next
item to be mixed and they would point to where in the value vec it
currently was.
Couldn't figure out how the pointers would work/what type they are. Are
they &i32 or Rc? or Rc::new(RefCell::Clone())?

Brute Force
Store the original order and value together.
When I need the ith value, I just search the whole vec for the right
element.