Day 20 Part 2
Rank 4283
But I cheated
I found a solution online. Once I read it, I realized this is how I was
trying to solve the problem with having a list of indicies and a list of
values.
This version changes the indices, not the values around. Which makes
sense.
And instead of swapping endlessly, it just does +1 and -1 to all the
indicies that need it.

when moving an element,
stuff to the right of where you started will all shift left -1
stuff to the right of where you ended gets shifted back +1
could just be if start < ix && ix < end => -1, right?
then set the ending point to what is should be.

rem_euclid
is basically my add_modd and sub_mod built in. Now I know how to use it.

finding the first, second, third is a bit odd
I kinda understand it. But haven't looked at it enough to explain it.