Day 21 Part 2
Rank 5127

I made a system to calculate the left half of the root. When you reached
humn in one of your children, add yourself to the cant pile. If your child
is a member of the can't pile, so are you. Go until the root is in the
cant pile. Check which of the root's children is in the cant pile and
which is in the values pile.
I didn't really know where to go from there.
I could go from the root and rearrange the tree to solve for value =
a+b*c/d+e-f*humn
I kind of know how I could do that, but not really sure how to code it.

What I actually did
Brute force
Guess a value for humn
evaluate
check if it's high or low
guess another value
repeat until it matches.
Wolf fencing
Pick an upper and lower value
guess the number in the middle.
Based on the results of the guess, replace the upper or lower.
I didn't really implement this. I just manually put in values and did it
myself. I was trying some example inputs to figure out what result would
result in replacing which value. But then I just kind of got in a rhythym
copy and pasting so just kept going instead of trying to automate it.
Whatever gets the answer I guess