Day 23 part 1
Made empty_direction work with the changing priority
loops 4 tmes to get the 4 orders
check the modulo to see if it is the right order.
loop i=0 will only match 1 of the directions
A lot of boilerplate code, not a lot of logic. making the 8 positions
takes a lot of space