Day 24 Cleanup, Day 25 Part 1
Works
Need to test a lot of different numbers
I had tests passing without being able to make numbers with a 0 in the
middle. And I was testing >10 numbers.

Checking vs the current power of 5 gives mostly right answers, but
actually need to test vs current power of 5 and all subsequent powers of 5
Had an abs in the wrong spot
Need to test if a non-zero value here would be worse than a 0. The
positive branch needs the abs bc it is the side that subtracts so it
creates the negative number that uses the negative branch.

Another bug was the max power
max power is 122
so 1*25 + 2*5 + 2*1
only 1x the largest power